refactored locations and types
