+++ model/signature/exception/SignatureException.php
@@ -20,7 +20,7 @@
 
 declare(strict_types=1);
 
-namespace oat\remoteProctoring\model\signature;
+namespace oat\remoteProctoring\model\signature\exception;
 
 use common_Exception;
