+++ test/unit/model/signature/Sha256SignatureTest.php
@@ -24,7 +24,7 @@ namespace oat\remoteProctoring\test\unit\model\signature;
 
 use oat\generis\test\TestCase;
 use oat\remoteProctoring\model\signature\Sha256Signature;
-use oat\remoteProctoring\model\signature\SignatureException;
+use oat\remoteProctoring\model\signature\exception\SignatureException;
 use Psr\Http\Message\RequestInterface;
 
 class Sha256SignatureTest extends TestCase
